Outsource the prediction in a separate interface and class
